requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
sentence-transformers
huggingface-hub